refactor: clean up pagination type and related services
