--- src/services/data.ts
@@ -1,9 +1,0 @@
-import { PaginationType } from "@components/Pagination";
-
-export function getPaginationProps(responseData: PaginationType): PaginationType {
-	return {
-		page: responseData.page,
-		pageSize: responseData.pageSize,
-		totalCount: responseData.totalCount,
-	};
-}
